+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+### Fixed
+
+- Pass languages to the TaxClassFormPart to show multilanguage forms
+
 ## [2.6.4] - 2021-04-19
 
 ### Changed