+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
+## [2.6.5] - 2021-04-22
+
 ### Changed
 
 - Front: consider the supplier from context or kwargs while reversing the product url